+++ cpp/Platform.Hashing/Platform.Hashing.Tuple.h
@@ -11,7 +11,7 @@ namespace Platform::Hashing
     template <class TupleType, size_t Index = std::tuple_size<TupleType>::value - 1>
     struct Tuple
     {
-        static void Combine(std::uint32_t& hashAccumulator, const TupleType& tuple)
+        static void Combine(std::size_t & hashAccumulator, const TupleType& tuple)
         {
             Tuple<TupleType, Index - 1>::Combine(hashAccumulator, tuple);
             std::size_t hash = Hashing::Hash(std::get<Index>(tuple));
@@ -22,7 +22,7 @@ namespace Platform::Hashing
     template <class TupleType>
     struct Tuple<TupleType, 0>
     {
-        static void Combine(std::uint32_t& hashAccumulator, const TupleType& tuple)
+        static void Combine(std::size_t& hashAccumulator, const TupleType& tuple)
         {
             std::size_t hash = Hashing::Hash(std::get<0>(tuple));
             Hashing::Combine(hashAccumulator, hash);
@@ -37,7 +37,7 @@ namespace std
     {
         std::size_t operator()(const std::tuple<Types...> &tuple) const
         {
-            std::uint32_t hash = 0;
+            std::size_t hash = 0;
             Platform::Hashing::Tuple<std::tuple<Types...>>::Combine(hash, tuple);
             return Platform::Hashing::Expand(hash);
         }
